requirements.txt
Django==1.10
djangorestframework==3.5.3
httpie==0.9.3
Pillow==3.1.1
psycopg2==2.6.1
Pygments==2.1
requests==2.9.1
